refactor: move secretprovider into seperate package
